Update graph_app_token.py
Removed the bearer mode - all that did was return a string "Bearer <token>". Now we just return the token by default or the entire raw response

Formatted code a little.

Runtime error for http responses other than 401 prints the response.json() as indented json to make it easier to read on screen